requirements.txt
discord.py>=2.3.0,<3.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
pywinpty>=2.0.0
